[Core] Move Profiling macros to Core project
